requirements.txt
PyQt6==6.6.1
python-docx==1.1.0
pdfplumber==0.10.4
PyMuPDF==1.23.22
reportlab==4.1.0
openai==1.14.3
keyring==24.3.1
cryptography==42.0.5
pandas==2.2.1
openpyxl==3.1.2
tqdm==4.66.2
requests==2.31.0
httpx==0.27.0
